# ch_4/api-docs.json
{
  "info": {
    "description": "Banking System",
    "title": "Swagger API",
    "version": "1.0.0"
  },
  "host": "localhost:undefined",
  "basePath": "/api/v1",
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "paths": {
...
    "/api/v1/accounts": {
      "post": {
        "parameters": [
          {
            "name": "bankAccount",
            "in": "body",
            "description": "Account data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/BankAccount"
            }
          },
          {
            "name": "bankAccount",
            "in": "body",
            "description": "User's ID (e.g., 1)",
            "required": false,
            "type": "number"
          },
          {
            "name": "bankAccount",
            "in": "body",
            "description": "User's bank name (e.g., \"Bank Mandiri\")",
            "required": false,
            "type": "string"
          },
          {
            "name": "bankAccount",
            "in": "body",
            "description": "User's bank account number (e.g., \"1234567890\")",
            "required": false,
            "type": "string"
          },
          {
            "name": "bankAccount",
            "in": "body",
            "description": "User's account balance (e.g., 10000)",
            "required": false,
            "type": "number"
          }
        ],
        "description": "",
        "tags": [
          "Accounts"
        ],
        "responses": {
          "201": {
            "description": "Successfully created bank account"
          },
          "400": {
            "description": "Invalid input data"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Create a new bank account"
      },
      "get": {
        "parameters": [],
        "description": "",
        "tags": [
          "Accounts"
        ],
        "responses": {
          "200": {
            "description": "List of accounts",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/BankAccount"
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get all accounts"
      }
    },
    "/api/v1/accounts/:accountId": {
      "get": {
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "description": "The ID of the account",
            "required": true,
            "type": "integer"
          }
        ],
        "description": "",
        "tags": [
          "Accounts"
        ],
        "responses": {
          "200": {
            "description": "Account Detail"
          },
          "404": {
            "description": "Account not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get account detail by Account ID"
      }
    },
    "/api/v1/accounts/deposit": {
      "post": {
        "parameters": [
          {
            "name": "deposit",
            "in": "body",
            "description": "Deposit data",
            "required": true,
            "type": "object"
          },
          {
            "name": "deposit",
            "in": "body",
            "description": "The ID of the account to deposit",
            "required": false,
            "type": "integer"
          },
          {
            "name": "deposit",
            "in": "body",
            "description": "The amount to deposit (harus lebih besar dari nol)",
            "required": false,
            "type": "number"
          }
        ],
        "description": "",
        "tags": [
          "Accounts"
        ],
        "responses": {
          "200": {
            "description": "Deposit Berhasil"
          },
          "400": {
            "description": "Jumlah deposit harus lebih besar dari nol"
          },
          "404": {
            "description": "Account not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Deposit money into an account"
      }
    },
    "/api/v1/accounts/withdraw": {
      "post": {
        "parameters": [
          {
            "name": "withdraw",
            "in": "body",
            "description": "Withdraw data",
            "required": true,
            "type": "object"
          },
          {
            "name": "withdraw",
            "in": "body",
            "description": "The ID of the account to withdraw",
            "required": false,
            "type": "integer"
          },
          {
            "name": "withdraw",
            "in": "body",
            "description": "The amount to withdraw",
            "required": false,
            "type": "number"
          }
        ],
        "description": "",
        "tags": [
          "Accounts"
        ],
        "responses": {
          "200": {
            "description": "Penarikan saldo berhasil"
          },
          "400": {
            "description": "Saldo tidak mencukupi"
          },
          "404": {
            "description": "Account not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Withdraw money into an account"
      }
    },
    "/api/v1/accounts/bank-account": {
      "get": {
        "parameters": [],
        "description": "",
        "tags": [
          "Accounts"
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Account not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Retrieve bank account details for the authenticated user"
      }
    },
...
    "/api/v1/transactions": {
      "post": {
        "parameters": [
          {
            "name": "transaction",
            "in": "body",
            "description": "Transaction data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Transaction"
            }
          },
          {
            "name": "transaction",
            "in": "body",
            "description": "The ID of the source account",
            "required": false,
            "type": "number"
          },
          {
            "name": "transaction",
            "in": "body",
            "description": "The ID of the destination account",
            "required": false,
            "type": "number"
          },
          {
            "name": "transaction",
            "in": "body",
            "description": "The amount to transfer (must be greater than zero)",
            "required": false,
            "type": "number"
          }
        ],
        "description": "",
        "tags": [
          "Transactions"
        ],
        "responses": {
          "201": {
            "description": "Successful transaction"
          },
          "400": {
            "description": "Saldo tidak mencukupi"
          },
          "404": {
            "description": "Akun sumber atau tujuan tidak ditemukan"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Transfer money from one account to another account\n/"
      },
      "get": {
        "parameters": [],
        "description": "",
        "tags": [
          "Transactions"
        ],
        "responses": {
          "200": {
            "description": "List of transactions",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Transaction"
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get all transactions\n/"
      }
    },
    "/api/v1/transactions/:transactionId": {
      "get": {
        "parameters": [
          {
            "name": "transactionId",
            "in": "path",
            "description": "The ID of the transaction",
            "required": true,
            "type": "integer"
          }
        ],
        "description": "",
        "tags": [
          "Transactions"
        ],
        "responses": {
          "200": {
            "description": "Transaction detail"
          },
          "404": {
            "description": "Transaction not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get transaction detail by Transaction ID\n/"
      }
    },
...
  },
...
  "securityDefinitions": {

  },
...
}